==> 1. Converting synsets to lexunits.ipynb <==
# 1. Converting synsets to lexunits

All pictographs are linked to a synset from the Cornetto database. However, later on, we will have to get examples from DutchSemCor, which works on the level of lexunits. Therefore, we will convert all synset ids to lexunit ids in this notebook.

## Loading the lexunit-synset bridge

First, we load in the lexunit-synset bridge. This will allow us to convert from synsets to lexunits.

from picto2vec.lexsynbridge import LexSynBridge

Then, we load in our conversion dataset. It might take a while to load in, since it internally converts this dataset to a Python `dict` for faster conversion. After loading, we test the conversion.

lexsynbridge = LexSynBridge("data/lex2syn_original.csv")

lexsynbridge.syn2lex("d_n-39469")

## Converting synsets to lexunits

That works! Now, let's load in our pictograph dataset and convert all columns with synsets in them to lexunits.

import pandas as pd
import numpy as np

sense_df = pd.read_csv("data/sclera.csv", names=[ "lemma", "synset", "relation", "head", "headrel", "dependent", "deprel", "antonym", "number", "lemma_english", "able" ])
sense_df.head()

sense_df['lexunit'] = sense_df.apply(lambda row: lexsynbridge.syn2lex(row["synset"]), axis=1)
sense_df['head_lexunit'] = sense_df.apply(lambda row: ",".join(list(filter(lambda lexunit: lexunit != False, list(map(lambda synset: lexsynbridge.syn2lex(synset), row["head"].split(",")))))) if type(row["head"]) == str else np.nan, axis=1)
sense_df['dependent_lexunit'] = sense_df.apply(lambda row: ",".join(list(filter(lambda lexunit: lexunit !=False, list(map(lambda synset: lexsynbridge.syn2lex(synset), row["dependent"].split(",")))))) if type(row["dependent"]) == str else np.nan, axis=1)
sense_df.head()

Some conversions will have failed. I don't know why this is the case, but it's not something I can fix, since I've noticed that most of the databases for Dutch are a bit... messy. We hope that this won't pose issues later on.

## Exporting the new dataset

We export the new dataset as JSON. This will be helpful later on, since pandas can be kinda slow when indexing.

sense_df.to_json("test_senses.json", orient="records")

That's it for this notebook!
