==> 2. Find sense references.ipynb <==
# 2. Find sense references

Now that we have converted all senses to lexunits, we can start finding these senses (actually lexunits, but I will use them interchangeably from now on) in the DutchSemCor corpus. This is what we'll be doing in this notebook.

There are quite a few constants we'll be reusing across notebooks, which is why these are defined in `constants.py`.

from constants import *

import os
import re
import pandas as pd
import json
from glob import glob

## Loading our dataset

First, we load in the dataset again we made in the previous notebook.

sense_df = pd.read_json("test_senses.json")

## Collecting all lexunits

Lexunits can come from various places. Either from a unary record, of from a record consisting of a head and a dependent. For completion's sake, we will collect examples of all lexunits in the dataset.

lexunits = []
for index, row in sense_df.iterrows():
    lexunits += row["lexunit"].split(",") if type(row["lexunit"]) == str else ""
    lexunits += row["head_lexunit"].split(",") if type(row["head_lexunit"]) == str else ""
    lexunits += row["dependent_lexunit"].split(",") if type(row["dependent_lexunit"]) == str else ""

lexunits = list(set(lexunits))    

token_find_re = "|".join(lexunits)
token_find_re

Now, we will use grep to find all lines in the DutchSemCor corpus pointing to corpus examples of senses we're interested in. The lines will be collected in ".match" files.  
Please **MOVE** these .match files to the path specified in DUTCHSEMCOR_FILTERED_CSV_PATH afterwards.

!for f in data/dutsemcor/*.xml ; do grep -E -i -w 'regex_here' "$f" -A9999 > "${f%.xml}.match" ; done

## Checking for dead references

Not all references in DutchSemCor are transparent. Therefore, we check for every single match whether we can actually find the files in our dataset. Lassy hits are provided by Vincent Vandeghinste.

lassy_hits = pd.read_csv(f"{DATA_PATH}/lassy_hits.tsv",
                         sep="\t",
                         names=["docid",
                                "xmlid",
                                "previous_sentence",
                                "current_sentence",
                                "next_sentence"],
                        header=None)

We create a list of all match files we found earlier.

xml_file_paths = glob("{}*.match".format(DUTCHSEMCOR_FILTERED_CSV_PATH))
xml_file_paths.sort()

Now, we define a function which will look for the appropriate corpus file. The reference's information is appended to the appropriate JSONL file (per sense). If the reference cannot be found, "lost" is set to `True`.

def find_senses_from_xml(xml_file_path):
    # Check if the file is empty
    if os.stat(xml_file_path).st_size == 0:
        return []
    
    with open(xml_file_path) as reader:
        xml_content = reader.read()
        
    found_tokens = xml_content.split("\n")
        
    # Create a dataframe row for every token we found
    for found_token in tqdm(found_tokens, leave=False):
        # Sometimes, there are erroneous lines included. If there is no match,
        # just skip to the next line
        token_attributes = re.search(r'lemma="(.*?)" pos="(.*?)" sense="(.*?)" token_id="(.*?)"', found_token)
        if token_attributes is None:
            continue
        
        token_attributes = token_attributes.groups()
        
        # token_id has the following structure:
        # WR-P-P-G-0000070187.p.13.s.5.w.25
        # WR-P-P-G-0000070187 = docid
        # p.13.s.5.w.25 = xmlid
        # We can only search for xmlid. In the corpus result, we can further filter
        # for docid. So we need both. We use regular expressions.
        token_id = token_attributes[3]
        doc_id, xml_id = re.search(r"(.*?)\.(.*)", token_id).groups()
        
        lemma = token_attributes[0].replace("&#195;&#168;", "è")
        pos = token_attributes[1]
        sense = token_attributes[2]
        
        lost = False
        
        if doc_id.startswith("CGN"):
            cgn_id = re.search(r"_(.*?)$", doc_id).groups()[0]
            plk_filename = f"{SONAR_PATH}{cgn_id}.plk"
                
            if not os.path.exists(plk_filename):
                lost = True
        elif doc_id.startswith("WR-P-P-I"):
            lassy_hit = lassy_hits.loc[(lassy_hits["docid"] == doc_id) & 
                                       (lassy_hits["xmlid"] == xml_id)]
            
            if len(lassy_hit) == 0:
                lost = True
        else:        
            xml_filename = f"{SONAR_PATH}{doc_id}.folia.xml"
        
            if not os.path.exists(xml_filename):
                lost = True
        
        row = { "lemma": lemma,
                "pos": pos,
                "sense": sense,
                "xmlid": xml_id,
                "docid": doc_id,
                "lost": lost }
        
        jsonl_path = f"sense_example_references/{row['sense']}.jsonl"
        with open(jsonl_path, "at") as writer:
            json_raw = json.dumps(row)
            writer.write(f"{json_raw}\n")

We apply the function for every .match file.

for xml_file_path in tqdm(xml_file_paths):
    find_senses_from_xml(xml_file_path)
